backend/requirements.txt
google-generativeai
python-dotenv
pypdf
chromadb
langchain
langchain-google-genai
langchain-community
langchain-chroma